Remove AnsiCast for 340
