+++ sql-plugin/src/main/320/scala/com/nvidia/spark/rapids/shims/SparkShims.scala
@@ -29,7 +29,8 @@ import org.apache.spark.sql.types.StructType
 
 object SparkShimImpl extends Spark320PlusShims
     with Spark320PlusNonDBShims
-    with Spark31Xuntil33XShims {
+    with Spark31Xuntil33XShims
+    with Spark320until340Shims{
   override def getSparkShimVersion: ShimVersion = ShimLoader.getShimVersion
 
   override def getFileScanRDD(
