+++ sql-plugin/src/main/321db/scala/com/nvidia/spark/rapids/shims/SparkShims.scala
@@ -38,7 +38,7 @@ import org.apache.spark.sql.rapids.GpuFileSourceScanExec
 import org.apache.spark.sql.rapids.shims.GpuFileScanRDD
 import org.apache.spark.sql.types._
 
-object SparkShimImpl extends Spark321PlusShims {
+object SparkShimImpl extends Spark321PlusShims with Spark320until340Shims {
 
   override def getSparkShimVersion: ShimVersion = ShimLoader.getShimVersion
