+++ sql-plugin/src/main/340/scala/com/nvidia/spark/rapids/SparkShims.scala
@@ -19,6 +19,7 @@ package com.nvidia.spark.rapids.shims
 import com.nvidia.spark.rapids._
 
 import org.apache.spark.rapids.shims.GpuShuffleExchangeExec
+import org.apache.spark.sql.catalyst.expressions.Expression
 import org.apache.spark.sql.catalyst.plans.physical.SinglePartition
 import org.apache.spark.sql.execution.{CollectLimitExec, GlobalLimitExec, SparkPlan}
 import org.apache.spark.sql.execution.exchange.ENSURE_REQUIREMENTS
@@ -71,5 +72,8 @@ object SparkShimImpl extends Spark330PlusShims {
 
   override def getExecs: Map[Class[_ <: SparkPlan], ExecRule[_ <: SparkPlan]] =
     super.getExecs ++ shimExecs
+
+  // AnsiCast is removed from Spark3.4.0
+  override def ansiCastRule: ExprRule[_ <: Expression] = null
 
 }
