+++ sql-plugin/src/main/scala/com/nvidia/spark/rapids/GpuParquetScan.scala
@@ -599,6 +599,9 @@ private case class GpuParquetFileFilterHandler(@transient sqlConf: SQLConf) exte
     }
   }
 
+  @scala.annotation.nowarn(
+    "msg=method readFooter in class ParquetFileReader is deprecated"
+  )
   def readAndSimpleFilterFooter(
       file: PartitionedFile,
       conf : Configuration,
